requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>= 8.2.0,<=8.3.0
djangorestframework-simplejwt>=4.7.2,<5.0.0
PyJWT>=2.0.0,<2.5.0
cryptography>=3.4.7
python-socketio[asgi]>=5.0
aiohttp==3.8.6
uvicorn==0.23.2
gunicorn==21.2.0
redis>=4.0.0
channels>=3.0.4,<4.0.0
pymongo==4.7.0
celery==5.2.7
numpy==1.23.5
scipy==1.9.3
pandas==1.5.3
scikit-learn==1.2.2
tqdm==4.65.0
torch>=1.12.0
sentence-transformers>=2.2.2
transformers>=4.35.0
pinecone>=3.0.0
django-celery-beat==2.5.0

